+++ src/main/java/com/maze/stulog/subscription/application/SubscriptionService.java
@@ -1,10 +1,10 @@
-package com.maze.stulog.schedule.application;
+package com.maze.stulog.subscription.application;
 
 import static com.maze.stulog.common.error.ExceptionCode.SUBSCRIPTION_NOT_FOUND;
 
 import com.maze.stulog.common.error.BusinessException;
 import com.maze.stulog.member.domain.Member;
-import com.maze.stulog.schedule.domain.Subscription;
+import com.maze.stulog.subscription.domain.Subscription;
 import com.maze.stulog.schedule.domain.repository.SubscriptionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
